feat(type-safety): enable strict mypy and add type annotations across codebase (#3)
* feat(type-safety): enable strict mypy and add type annotations across codebase

- Re-enable strict mypy for src/synapse24 (removed exclude, enabled disallow_untyped_defs)
- Add py.typed marker for PEP 561 compliance
- Fix type annotations in signal_quality modules (ppg, ecg, base)
- Fix type annotations in ingestion modules (wesad, mitbih)
- Fix type annotations in utils/xdf.py (LSLStreamManager, XDF writer helpers)
- Remove ingestion from coverage omit list (now tested)
- Add comprehensive test coverage for ingestion pipelines (86% total coverage)
- All tests passing (78 passed, 4 skipped), mypy clean, ruff clean

* ci: fix mypy invocation to use --package synapse24 with config file

* test: fix open mock for download_wesad test (patch builtins.open)

* test: fix download_wesad test with proper mock parameter ordering

+++ src/synapse24/signal_quality/base.py
@@ -342,4 +342,4 @@ def compute_snr(signal: np.ndarray, noise: np.ndarray) -> float:
     noise_power = np.mean(noise**2)
     if noise_power == 0:
         return float("inf")
-    return 10 * np.log10(signal_power / noise_power)
+    return float(10 * np.log10(signal_power / noise_power))
